fix: regression introduced due to refactoring
Porting the application to 64 bits introduced a regression,
where a seemingly innocent change from addresses to indices
caused the value `unused`, previously zero, to be used.

Before the refactoring, `unused` represented a null pointer,
and was unequal to any other address value.

After the refactoring, `unused` becomes equal to the 0th index,
and results in a bug that prevents the backward-iteration
to stop at the 0th node (because its parent is `0` the code thinks
the it is unused), so it continues backward resulting in an array
out-of-bounds error.

To fix this one could simply define an `unused` value that is
not used by any index. I changed `Row` from `uint32_t` to `int32_t`
and set `unused` to be `-1`.

As an effect, this effectively halves the max usable memory from
16GB (4 bytes * 2^32) to 8GB (4 bytes * 2^31), but overall I think
it is worth preventing any further regressions caused by a change
from an unsigned integer to a signed integer.